cleaned up some files and renamed testCase to reactTestCase
