+++ src/containers/LeftPanel/TestMenu/TestMenu.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import styles from '../TestMenu/TestMenu.module.scss';
-import { addAction, addAssertion, addRender } from '../../../context/testCaseActions';
+import { addAction, addAssertion, addRender } from '../../../context/reactTestCaseActions';
 import NewTestModal from '../../NavBar/Modals/NewTestModal';
 
 const TestMenu = ({ dispatchToTestCase, dispatchToMockData }) => {
