+++ src/context/reactTestCaseReducer.js
@@ -1,0 +1,260 @@
+import { createContext } from 'react';
+import { actionTypes } from './reactTestCaseActions';
+
+export const ReactTestCaseContext = createContext(null);
+
+export const reactTestCaseState = {
+  testStatement: '',
+  hasReact: 0,
+  statements: [
+    {
+      id: 0,
+      type: 'render',
+      componentName: '',
+      filePath: '',
+      props: [],
+      hasProp: false,
+    },
+    {
+      id: 1,
+      type: 'assertion',
+      queryVariant: '',
+      querySelector: '',
+      queryValue: '',
+      isNot: false,
+      matcherType: '',
+      matcherValue: '',
+      suggestions: [],
+    },
+  ],
+};
+
+let statementId = 2;
+let renderPropsId = 0;
+
+const createAction = () => ({
+  id: statementId++,
+  type: 'action',
+  eventType: '',
+  eventValue: null,
+  queryVariant: '',
+  querySelector: '',
+  queryValue: '',
+  suggestions: [],
+});
+
+const createAssertion = () => ({
+  id: statementId++,
+  type: 'assertion',
+  queryVariant: '',
+  querySelector: '',
+  queryValue: '',
+  isNot: false,
+  matcherType: '',
+  matcherValue: '',
+  suggestions: [],
+});
+
+const createRerender = (componentName, filePath) => ({
+  id: statementId++,
+  type: 'render',
+  componentName,
+  filePath,
+  props: [],
+});
+
+const createRenderProp = () => ({
+  id: renderPropsId++,
+  propKey: '',
+  propValue: '',
+});
+
+export const reactTestCaseReducer = (state, action) => {
+  Object.freeze(state);
+  let statements = [...state.statements];
+  let lastAssertionStatement;
+
+  switch (action.type) {
+    case actionTypes.TOGGLE_REACT:
+      return {
+        ...state,
+        hasReact: state.hasReact + 1,
+      };
+    case actionTypes.UPDATE_STATEMENTS_ORDER:
+      const firstRenderStatement = statements[0];
+      lastAssertionStatement = statements[statements.length - 1];
+      statements = [firstRenderStatement, ...action.draggableStatements, lastAssertionStatement];
+      return {
+        ...state,
+        statements,
+      };
+    case actionTypes.UPDATE_TEST_STATEMENT:
+      let testStatement = action.testStatement;
+      return {
+        ...state,
+        testStatement,
+      };
+    case actionTypes.ADD_ACTION:
+      lastAssertionStatement = statements.pop();
+      statements.push(createAction(), lastAssertionStatement);
+      return {
+        ...state,
+        statements,
+      };
+    case actionTypes.DELETE_ACTION:
+      lastAssertionStatement = statements.pop();
+      statements = statements.filter(statement => statement.id !== action.id);
+      statements.push(lastAssertionStatement);
+      return {
+        ...state,
+        statements,
+      };
+    case actionTypes.UPDATE_ACTION:
+      statements = statements.map(statement => {
+        if (statement.id === action.id) {
+          statement.eventType = action.eventType;
+          statement.eventValue = action.eventValue;
+          statement.queryVariant = action.queryVariant;
+          statement.querySelector = action.querySelector;
+          statement.queryValue = action.queryValue;
+          statement.suggestions = action.suggestions;
+        }
+        return statement;
+      });
+      return {
+        ...state,
+        statements,
+      };
+    case actionTypes.ADD_ASSERTION:
+      lastAssertionStatement = statements.pop();
+      statements.push(createAssertion(), lastAssertionStatement);
+      return {
+        ...state,
+        statements,
+      };
+    case actionTypes.DELETE_ASSERTION:
+      lastAssertionStatement = statements.pop();
+      statements = statements.filter(statement => statement.id !== action.id);
+      statements.push(lastAssertionStatement);
+      return {
+        ...state,
+        statements,
+      };
+    case actionTypes.UPDATE_ASSERTION:
+      statements = statements.map(statement => {
+        if (statement.id === action.id) {
+          statement.queryVariant = action.queryVariant;
+          statement.querySelector = action.querySelector;
+          statement.queryValue = action.queryValue;
+          statement.isNot = action.isNot;
+          statement.matcherType = action.matcherType;
+          statement.matcherValue = action.matcherValue;
+          statement.suggestions = action.suggestions;
+        }
+        return statement;
+      });
+      return {
+        ...state,
+        statements,
+      };
+    case actionTypes.ADD_RENDER:
+      lastAssertionStatement = statements.pop();
+      const renderComponentName = state.statements[0].componentName;
+      const renderFilePath = state.statements[0].filePath;
+      statements.push(createRerender(renderComponentName, renderFilePath), lastAssertionStatement);
+      return {
+        ...state,
+        statements,
+      };
+    case actionTypes.DELETE_RENDER:
+      lastAssertionStatement = statements.pop();
+      statements = statements.filter(statement => statement.id !== action.id);
+      statements.push(lastAssertionStatement);
+      return {
+        ...state,
+        statements,
+        lastAssertionStatement,
+      };
+    case actionTypes.UPDATE_RENDER_COMPONENT:
+      statements = statements.map(statement => {
+        if (statement.type === 'render') {
+          statement.componentName = action.componentName;
+          statement.filePath = action.filePath;
+        }
+        return statement;
+      });
+      return {
+        ...state,
+        statements,
+      };
+    case actionTypes.ADD_RENDER_PROP:
+      statements = statements.map(statement => {
+        if (statement.id === action.renderId) {
+          statement.props.push(createRenderProp());
+        }
+        return statement;
+      });
+      return {
+        ...state,
+        statements,
+        hasProp: !statements[0].hasProp,
+      };
+    case actionTypes.DELETE_RENDER_PROP:
+      statements = statements.map(statement => {
+        if (statement.id === action.renderId) {
+          statement.props = statement.props.filter(prop => prop.id !== action.propId);
+        }
+        return statement;
+      });
+      return {
+        ...state,
+        statements,
+      };
+    case actionTypes.UPDATE_RENDER_PROP:
+      statements = statements.map(statement => {
+        if (statement.id === action.renderId) {
+          statement.props.map(prop => {
+            if (prop.id === action.propId) {
+              prop.propKey = action.propKey;
+              prop.propValue = action.propValue;
+            }
+            return prop;
+          });
+        }
+        return statement;
+      });
+      return {
+        ...state,
+        statements,
+      };
+
+    case actionTypes.CREATE_NEW_TEST:
+      return {
+        hasReact: 0,
+        testStatement: '',
+        statements: [
+          {
+            id: 0,
+            type: 'render',
+            componentName: '',
+            filePath: '',
+            props: [],
+            hasProp: false,
+          },
+          {
+            id: 1,
+            type: 'assertion',
+            queryVariant: '',
+            querySelector: '',
+            queryValue: '',
+            isNot: false,
+            matcherType: '',
+            matcherValue: '',
+            suggestions: [],
+          },
+        ],
+      };
+    default:
+      return state;
+  }
+};
